Upadted UI

+++ packetsimulator-master/packetsimulator/src/main/java/com/network/packetsimulator/graph/Edge.java
@@ -13,7 +13,7 @@ public class Edge {
         return to;
     }
 
-    public int getWeight() {
-        return (int) weight;
+    public double getWeight() {
+        return weight;
     }
 }
